pyproject.toml
[project]
name = "pythonbddframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.0",
    "playwright>=1.55.0",
    "pytest>=8.4.2",
    "pytest-bdd>=8.1.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.8.0",
]
